pyproject.toml
[project]
name = "costoptimizer"
version = "0.1.0"
description = "AI-driven Multi-Cloud Cost Optimization SaaS Platform"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "email-validator>=2.1.0",
    "bcrypt==4.1.3",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
# Install cloud SDKs only for the providers you need
aws = ["boto3>=1.34.0"]
azure = [
    "azure-identity>=1.15.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-compute>=30.4.0",
    "azure-mgmt-monitor>=6.0.2",
    "azure-mgmt-resource>=23.0.1",
    "azure-mgmt-storage>=21.1.0",
]
gcp = [
    "google-cloud-billing>=1.13.0",
    "google-cloud-compute>=1.17.0",
    "google-cloud-monitoring>=2.19.0",
    "google-cloud-bigquery>=3.17.0",
    "google-cloud-resource-manager>=1.12.0",
]
llm = ["anthropic>=0.40.0"]
finetune = [
    "peft>=0.7.0",
    "trl>=0.7.0",
    "transformers>=4.40.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.18.0",
    "accelerate>=0.30.0",
]
postgres = ["asyncpg>=0.29.0", "psycopg2-binary>=2.9.9"]
workers = ["redis>=5.0.1", "celery[redis]>=5.3.6"]
all = ["costoptimizer[aws,azure,gcp,llm,finetune,postgres,workers]"]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
